修复bug

+++ gaussiansplatting/gaussian_renderer/__init__.py
@@ -80,7 +80,8 @@ def render(
         bg=bg_color,
         scale_modifier=scaling_modifier,
         viewmatrix=viewpoint_camera.world_view_transform,
-        projmatrix=viewpoint_camera.full_proj_transform,
+        # 把BFloat16转换成float32
+        projmatrix=viewpoint_camera.full_proj_transform.float(),
         sh_degree=pc.active_sh_degree,
         campos=viewpoint_camera.camera_center,
         prefiltered=False,
